Move frint-cli source int src folder
