+++ packages/frint-cli/package.json
@@ -2,7 +2,6 @@
   "name": "frint-cli",
   "version": "5.2.0",
   "description": "CLI for Frint",
-  "main": "lib/index.js",
   "homepage": "https://github.com/frintjs/frint/tree/master/packages/frint-cli",
   "bin": {
     "frint": "./dist/bin/frint.js"
@@ -10,7 +9,7 @@
   "scripts": {
     "lint": "cross-env ./node_modules/.bin/tslint -p .",
     "transpile": "cross-env ./node_modules/.bin/tsc",
-    "test": "cross-env ../../node_modules/.bin/mocha --require ts-node/register --colors --recursive ./**/*.spec.ts",
+    "test": "cross-env ../../node_modules/.bin/mocha --require ts-node/register --colors --recursive ./src/**/*.spec.ts",
     "cover:run": "echo \"Nothing to cover\"",
     "cover:report": "echo \"Nothing to report\"",
     "cover": "npm run cover:run && npm run cover:report",
