+++ packages/frint-cli/src/commands/help.ts
