+++ packages/frint-cli/src/commands/init.ts
