+++ packages/frint-cli/src/commands/new.spec.ts
@@ -2,15 +2,15 @@ import { expect } from 'chai';
 import { App } from 'frint';
 import 'mocha';
 
-import createRootApp from '../root/index.mock';
-import InitCommand from './init';
+import createRootApp from '../index.mock';
+import NewCommand from './new';
 
-describe('frint-cli › commands › init', () => {
+describe('frint-cli › commands › new', () => {
   it('is a Frint App', () => {
     const RootApp = createRootApp();
     const rootApp = new RootApp();
-    rootApp.registerApp(InitCommand);
-    const commandApp = rootApp.getAppInstance('init');
+    rootApp.registerApp(NewCommand);
+    const commandApp = rootApp.getAppInstance('new');
 
     expect(commandApp).to.be.an.instanceOf(App);
   });
