+++ packages/frint-cli/src/commands/new.ts
