+++ packages/frint-cli/src/commands/version.ts
