+++ packages/frint-cli/src/index.mock.ts
